App Config: Unused/central dependencies, side-menu
Opted to use $state instead of the $location dependency for cleaner
implementations (not to mention we have $state anyway). Also, we added an
onEnter and onExit function for the app.gallery state that it
automatically disables the sideMenu when the user enters the gallery, and
enables it right back when the user leaves the gallery.