AppCtrl: $state vs $location, bug-fix on about modal
Utilizing $state instead of $location service for cleaner implementations.
Also fixed the bug on modal where if we click on the license, it does not
load the respective url on the system browser, which was fixed by adding a
scope to the $ionicPopup service. Also removed the $scope.isOnlineNotify
from our AppCtrl, opting to check for internet connections in our services
so that Controllers have to worry less about logic and more about just
transmitting data from view and our services to each other.