Auth: Added settings feature and pull of user id on login
Added settings feature that is stored on localStorage and pulled in for
that particular user using our application. We also pull their user id
from the server on login and use that id as one of the ways we can
identify the user so that when we are traversing over the gallery images,
if a user has rated a certain image or uploaded it him/herself, he/she
cannot rate it.