ImgNormalizer and other CSS image classes
ImgNormalizer will help us get images "normalized" for every phone screen
by expanding to smallest dimension of an image (whether it be by width or
height) to 100%, which will cause an overflow in the other side of the
image. To solve this, we use the solution provided at
http://stackoverflow.com/questions/14562457/center-oversized-image-in-div.
Other image css classes like responsive and block are also provided.