Moved dependencies to Bower with Gulp integration
Moved all of our dependencies to be handled by Bower to keep them
disconnected and easily upgradable. We also are using a copy of ngCordova
forked from the official ionic ng-cordova to
https://github.com/alastairparagas/ng-cordova. We made some important
changes on the File plugin with the addition of readFileAbsolute and
readFileMetadataAbsolute. Whenever ng-cordova upgrades, make sure to merge
the alastairparagas repo with the main ionic/ng-cordova repo.