Images: new compression method
Before, we used to use HTML5 canvas to try to degrade the quality of the
image and get the upload speeds up, but this was problematic and slow. To
solve this problem, we now take the metadata of the first picture taken by
the user and compare it against the standard size of the file we want sent
from the user's phone to our serve, which allows us to degrade the quality
of the image whenever the camera opens (using the Camera Options of
Cordova) to that produced quality value.